Refactor Automaton simulation
